requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
dhooks==1.1.4
idna==3.3
multidict==5.2.0
nextcord==2.0.0a3
PyYAML==6.0
requests==2.26.0
sqlitedict==1.7.0
typing-extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
